Separate Trellis resource matcher and request filter
Resolves #203
