+++ core/http/src/test/java/org/trellisldp/http/LdpUserResourceTest.java
@@ -41,6 +41,7 @@ public class LdpUserResourceTest extends AbstractLdpResourceTest {
         config.register(new AgentAuthorizationFilter(mockAgentService));
         config.register(new CacheControlFilter(86400, true, false));
         config.register(new WebSubHeaderFilter(HUB));
+        config.register(new TrellisHttpFilter());
         config.register(new CrossOriginResourceSharingFilter(asList(origin), asList("PATCH", "POST", "PUT"),
                         asList("Link", "Content-Type", "Accept-Datetime", "Accept"),
                         asList("Link", "Content-Type", "Memento-Datetime"), true, 100));
